requirements.txt
azure-core==1.31.0
azure-ai-services==0.1.0
azureml
azure-ai-documentintelligence
azure-ai-formrecognizer
